interfaces in business
Alvast even dit correct doen, zodat ik niet te veel in 1 pull request merge
