+++ InEenNotendop/HighscoreList.xaml.cs
@@ -9,6 +9,7 @@ using System.Windows.Controls;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
+using InEenNotendop.Business;
 using InEenNotendop.Data;
 
 namespace InEenNotendop.UI
