+++ InEenNotendop/SelectingWindow.xaml.cs
@@ -82,8 +82,7 @@ namespace InEenNotendop.UI
 
         private void FillDataGrid(int songId)
         {
-            HighScoresGrid.ItemsSource = _sqlDataAccess.GetDataForGrid(songId, 5);
-
+            HighScoresGrid.ItemsSource = _sqlDataAccess.GetTopScores(songId, 5);
         }
     }
 }
